+++ src/smc_lammps/generate/__init__.py
@@ -1,0 +1,3 @@
+from smc_lammps.generate import default_parameters, util
+
+__all__ = ["default_parameters", "util"]