+++ src/smc_lammps/generate/default_parameters.py
@@ -66,6 +66,15 @@ class Parameters:
     dna_config: str = "folded"
     "Initial DNA configuration to generate."
 
+    extension: float = 1.0
+    """The relative DNA extension: ratio of end-to-end distance to contour length.
+
+    This must be a value in the range (0.0, 1.0].
+
+    .. Note::
+        This should only be used with ``dna_config = "line"``.
+    """
+
     add_stopper_bead: bool = False
     "Add a bead that prevents the SMC from slipping off of the wrong end of the DNA."
 
@@ -93,8 +102,13 @@ class Parameters:
     spaced_beads_custom_stiffness: float = 1.0
     "Multiple of the default DNA stiffness."
 
-    spaced_beads_type: int = 1
-    "Type of beads to use, ``0`` -> fene/expand bonds, ``1`` -> rigid molecules."
+    spaced_beads_type: int = 2
+    """ Type of beads to use.
+
+        - ``0`` -- beads are interspersed between DNA segments with fene/expand bonds
+        - ``1`` -- beads are placed on top of the DNA forming rigid molecules (WARNING: likely to cause crashes)
+        - ``2`` -- beads are placed on top of the DNA and attached with harmonic bonds (default)
+    """
 
     ##################### Geometry #####################