+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.3] - 2026-06-15
+
+### Added
+
+- The DNA indices where spaced beads have been inserted are now saved to `post_processing_parameters.py`.
+- New `extension` parameter controls the relative extension of the DNA (for `Line` configuration only).
+- New `spaced_beads_type = 2` which uses harmonic bonds instead of a rigid molecule to hold beads in place.
+- The colors of SMC parts in VMD are now set consistently based on the type id.
+- New `bridge` option for `--visualize-follow` flag.
+
+### Fixed
+
+- Fixed error when setting `force = None` for `Line` DNA.
+- The `force` value now works at the DNA end points when setting `add_stopper_bead = True`.
+
 ## [0.2.2] - 2026-04-08
 
 ### Changed